+++ src/compas_dem/problem/__init__.py
@@ -1,0 +1,5 @@
+from .boundary_conditions import BoundaryConditions
+from .problem import Problem
+from .solvers import Solver
+
+__all__ = ["BoundaryConditions", "Problem", "Solver"]